Changing page_{start,end} to be 1-based
With page_articles giving the list of articles to display on the page,
there is no need for page_start and page_end to be indexes into the
article list. They are more useful for display purposes, such as showing
"Items X to Y of Z" in which case they should be 1-based.