Create summary from source instead of output
The old way of generating an article summary by splitting at an empty
line could result in unclosed html tags. This patch avoids that for
Markdown source files, by splitting on the source file instead of the
rendered output, and then rendering the summary separately. It also
provides a means to override the default summary generator, if a more
complex algorithm is required.