Add support for environments
As of https://github.com/middleman/middleman/pull/1293/files middleman
splits the notion of `:mode` and `:config`. This PR updates
middleman-blog to also handle this distinction.

This allows generation of blogposts using config for a particular
environment using the `--environment` flag. e.g:

`middleman article 'Awesome new post' --environment blog would look for
config within the `configure :blog do ...` block.

The main gotcha I see is that this requires `Time.zone` to be set in
config.rb also. I wonder though if that is related to one of the various
other date-related issues:
https://github.com/middleman/middleman-blog/issues/238,
https://github.com/middleman/middleman-blog/issues/191,
https://github.com/middleman/middleman-blog/issues/143