Only set ActiveSupport `to_time_preserves_timezone` on 8.0 and older
In ActiveSupport 8.1, the `to_time_preserves_timezone` setting was
deprecated and no longer has any effect on the implementation of
`to_time`, however it logs a deprecation warning when the extension is
initialised:

    DEPRECATION WARNING: `config.active_support.to_time_preserves_timezone`
    is deprecated and will be removed in Rails 8.2 (called from
    Middleman::BlogExtension#after_configuration at ...)

From 8.2, the setting will be removed entirely so to silence the warning
and ensure future compatibility, only set the mode on older versions of
ActiveSupport where it's needed.

Once middleman-core requires 8.1 or higher, this can be safely removed
again.